+++ src/Linna/Cache/ActionMultipleTrait.php
@@ -25,61 +25,58 @@ trait ActionMultipleTrait
      *
      * @param string $key     The unique key of this item in the cache.
      * @param mixed  $default Default value to return if the key does not exist.
+     *
+     * @return mixed The value of the item from the cache, or $default in case of cache miss.
      */
-    abstract public function get($key, $default = null);
+    abstract public function get(string $key, $default = null);
 
     /**
      * Express Requirements by Abstract Methods.
      *
      * @param string                $key   The key of the item to store.
      * @param mixed                 $value The value of the item to store, must be serializable.
-     * @param null|int|DateInterval $ttl   Optional. The TTL value of this item. If no value is sent and
-     *                                     the driver supports TTL then the library may set a default value
-     *                                     for it or let the driver take care of that.
+     * @param int                   $ttl   Optional. The TTL (time to live) value in seconds of this item.
+     *                                     If no value is sent and the driver supports TTL then the
+     *                                     library may set a default value for it or let the driver take care of that.
+     *
+     * @return bool True on success and false on failure.
      */
-    abstract public function set($key, $value, $ttl);
+    abstract public function set(string $key, $value, int $ttl = 0) : bool;
 
     /**
      * Express Requirements by Abstract Methods.
      *
      * @param string $key The cache item key.
      */
-    abstract public function has($key);
+    abstract public function has(string $key) : bool;
 
     /**
      * Express Requirements by Abstract Methods.
      *
      * @param string $key The unique cache key of the item to delete.
      */
-    abstract public function delete($key);
+    abstract public function delete(string $key) : bool;
 
     /**
      * {@inheritdoc}
      */
-    public function getMultiple($keys, $default = null)
+    public function getMultiple(array $keys, $default = null) : array
     {
-        if (!is_array($keys) && !($keys instanceof Traversable)) {
-            throw new InvalidArgumentException();
+        $arrayResult = [];
+        
+        foreach ($keys as $key) {
+            $arrayResult[$key] = $this->get($key, $default);
         }
 
-        $result = [];
-        foreach ((array) $keys as $key) {
-            $result[$key] = $this->get($key, $default);
-        }
-
-        return $result;
+        return $arrayResult;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function setMultiple($values, $ttl = null)
+    public function setMultiple(array $values, int $ttl = 0) : bool
     {
-        if (!is_array($values) && !($values instanceof Traversable)) {
-            throw new InvalidArgumentException();
-        }
-
-        foreach ((array) $values as $key => $value) {
+        foreach ($values as $key => $value) {
             $this->set($key, $value, $ttl);
         }
 
@@ -89,13 +86,9 @@ trait ActionMultipleTrait
     /**
      * {@inheritdoc}
      */
-    public function deleteMultiple($keys)
+    public function deleteMultiple(array $keys) : bool
     {
-        if (!is_array($keys) && !($keys instanceof Traversable)) {
-            throw new InvalidArgumentException();
-        }
-
-        foreach ((array) $keys as $key) {
+        foreach ($keys as $key) {
             $this->delete($key);
         }